backend/requirements.txt
# Core Flask
Flask==2.3.3
flask-cors==5.0.1
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Flask-Bcrypt==1.0.1

# Base de datos
SQLAlchemy==2.0.23
alembic==1.15.2
psycopg2-binary==2.9.9

# Utilidades
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.32.3
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0

# Para funcionalidades futuras
PyMuPDF==1.25.5
marshmallow==3.20.1
marshmallow-sqlalchemy==1.1.0

# Validación y serialización
marshmallow==3.20.1
marshmallow-sqlalchemy==1.1.0

# Utilidades
python-dotenv==1.0.0
PyJWT==2.8.0
itsdangerous==2.2.0

# Para funcionalidades futuras
PyMuPDF==1.25.5
requests==2.32.3

# Almacenamiento en caché
redis==4.5.5

# Procesamiento asíncrono
celery==5.3.6

# Desarrollo
gunicorn==21.2.0
pytest==7.4.0
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
tenacity==8.2.3
python-json-logger==2.0.7

# Documentación
apispec==6.3.0
apispec-webframeworks==0.5.0
flask-swagger-ui==4.11.1

# Observabilidad
sentry-sdk[flask]==1.45.0
prometheus-flask-exporter==0.22.4

# SDK de Anthropic
anthropic==0.51.0
